requirements.txt
fastapi
uvicorn
pydantic
PyJWT
python-dotenv
google-auth
requests
pytest